# vectispire-angular/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ComplianceEvaluation" : {
        "properties" : {
          "controls" : {
            "items" : {
              "$ref" : "#/components/schemas/ControlAssessment"
            },
            "type" : "array"
          },
          "framework" : {
            "enum" : [ "NIS_2", "ISO_27001", "EU_CRA", "DORA", "PCI_DSS", "SOC_2" ],
            "type" : "string"
          },
          "overallStatus" : {
            "enum" : [ "COMPLIANT", "PARTIAL", "NON_COMPLIANT" ],
            "type" : "string"
          },
          "scorePercentage" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "ComplianceSummary" : {
        "properties" : {
          "dueSoonCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "evaluations" : {
            "items" : {
              "$ref" : "#/components/schemas/ComplianceEvaluation"
            },
            "type" : "array"
          },
          "mttr" : {
            "$ref" : "#/components/schemas/MttrResult"
          },
          "overdueCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "passingGateTargets" : {
            "format" : "int32",
            "type" : "integer"
          },
          "targets" : {
            "items" : {
              "$ref" : "#/components/schemas/TargetCompliance"
            },
            "type" : "array"
          },
          "totalMonitoredTargets" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "ContainerCreateRequest" : {
        "properties" : {
          "image_name" : {
            "type" : "string"
          },
          "registry" : {
            "type" : "string"
          },
          "required_agent_label" : {
            "type" : "string"
          },
          "scanCron" : {
            "type" : "string"
          },
          "scanIntervalMinutes" : {
            "format" : "int32",
            "type" : "integer"
          },
          "tag" : {
            "type" : "string"
          },
          "tier" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ContainerSummary" : {
        "properties" : {
          "displayName" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "imageName" : {
            "type" : "string"
          },
          "lastScan" : {
            "$ref" : "#/components/schemas/LastScan"
          },
          "lastScheduledScanAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "openIssues" : {
            "format" : "int64",
            "type" : "integer"
          },
          "reference" : {
            "type" : "string"
          },
          "registry" : {
            "type" : "string"
          },
          "requiredAgentLabel" : {
            "type" : "string"
          },
          "scanCron" : {
            "type" : "string"
          },
          "scanIntervalMinutes" : {
            "format" : "int32",
            "type" : "integer"
          },
          "tag" : {
            "type" : "string"
          },
          "tier" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "LastScan" : {
        "properties" : {
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "error" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "status" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "MttrResult" : {
        "properties" : {
          "mttrBySeverityDays" : {
            "additionalProperties" : {
              "format" : "double",
              "type" : "number"
            },
            "type" : "object"
          },
          "overallMttrDays" : {
            "format" : "double",
            "type" : "number"
          },
          "resolvedCount" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "QueuedScan" : {
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "status" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ScopeCoverage" : {
        "properties" : {
          "declaredAssets" : {
            "format" : "int32",
            "type" : "integer"
          },
          "inScope" : {
            "format" : "int32",
            "type" : "integer"
          },
          "neverScanned" : {
            "format" : "int32",
            "type" : "integer"
          },
          "scannedRecently" : {
            "format" : "int32",
            "type" : "integer"
          },
          "stale" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "ScopeView" : {
        "properties" : {
          "coverage" : {
            "$ref" : "#/components/schemas/ScopeCoverage"
          },
          "statement" : {
            "type" : "string"
          },
          "targets" : {
            "items" : {
              "$ref" : "#/components/schemas/TargetRef"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "SecurityScorecard" : {
        "properties" : {
          "grade" : {
            "enum" : [ "A_PLUS", "A", "B", "C", "D", "F" ],
            "type" : "string"
          },
          "hasAttestation" : {
            "type" : "boolean"
          },
          "licenseViolationCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "openCriticalCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "openHighCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "openKevCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "overdueCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "recommendations" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "score" : {
            "format" : "int32",
            "type" : "integer"
          },
          "targetId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "targetKind" : {
            "type" : "string"
          },
          "targetName" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "TargetRef" : {
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "kind" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "TeamTargetAssignment" : {
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "kind" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "UserTargetAssignment" : {
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "kind" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "AgentKeyAuth" : {
        "description" : "Pre-shared authentication key for remote scanning agents",
        "in" : "header",
        "name" : "X-Agent-Key",
        "type" : "apiKey"
      },
      "ApiKeyAuth" : {
        "description" : "Programmatic API Key for CI/CD and automation integrations",
        "in" : "header",
        "name" : "X-API-Key",
        "type" : "apiKey"
      },
      "BearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "JWT Session Bearer token obtained via /api/v1/auth/login",
        "name" : "BearerAuth",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "contact" : {
      "name" : "Asmolabs Security Team",
      "url" : "https://github.com/asmolabs/vectispire"
    },
    "description" : "Automated supply-chain security, SBOM analysis, vulnerability triage, attack surface inventory, and compliance governance platform.",
    "license" : {
      "name" : "Apache-2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "title" : "Vectispire Control Plane REST API",
    "version" : "4.1.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/compliance/scope/containers/{id}" : {
      "put" : {
        "description" : "Marks one container image as inside or outside the certified scope.",
        "operationId" : "setContainer",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "in_scope",
          "required" : true,
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ScopeView"
                }
              }
            },
            "description" : "Membership set"
          }
        },
        "summary" : "Set an image's scope membership",
        "tags" : [ "Certified scope" ]
      }
    },
...
    "/api/v1/compliance/summary" : {
      "get" : {
        "description" : "Returns compliance scores across all regulatory frameworks.",
        "operationId" : "summary",
        "parameters" : [ {
          "description" : "Optional target ID filter",
          "in" : "query",
          "name" : "targetId",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ComplianceSummary"
                }
              }
            },
            "description" : "Compliance summary evaluated successfully"
          }
        },
        "summary" : "Get compliance summary",
        "tags" : [ "Compliance" ]
      }
    },
    "/api/v1/containers" : {
      "get" : {
        "operationId" : "list_7",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ContainerSummary"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "containers-controller" ]
      },
      "post" : {
        "operationId" : "create_5",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ContainerCreateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContainerSummary"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "containers-controller" ]
      }
    },
    "/api/v1/containers/{id}" : {
      "delete" : {
        "operationId" : "remove_3",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "containers-controller" ]
      },
      "patch" : {
        "operationId" : "update_3",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ContainerCreateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContainerSummary"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "containers-controller" ]
      }
    },
    "/api/v1/containers/{id}/scan" : {
      "post" : {
        "operationId" : "triggerScan_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/QueuedScan"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "containers-controller" ]
      }
    },
...
    "/api/v1/repositories/{id}/scan" : {
      "post" : {
        "description" : "Enqueues an immediate full security scan (SBOM, CVE, Secrets, SAST, IaC, APIs).",
        "operationId" : "triggerScan",
        "parameters" : [ {
          "description" : "Repository identifier",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/QueuedScan"
                }
              }
            },
            "description" : "Scan queued successfully"
          }
        },
        "summary" : "Trigger repository scan",
        "tags" : [ "Repositories" ]
      }
    },
...
    "/api/v1/scorecards/containers/{containerId}" : {
      "get" : {
        "description" : "Calculates security grade and risk posture for a container image.",
        "operationId" : "getContainerScorecard",
        "parameters" : [ {
          "description" : "Container ID",
          "in" : "path",
          "name" : "containerId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SecurityScorecard"
                }
              }
            },
            "description" : "Container scorecard retrieved successfully"
          }
        },
        "summary" : "Get container scorecard",
        "tags" : [ "Scorecards" ]
      }
    },
...
    "/api/v1/teams/{id}/targets" : {
      "get" : {
        "operationId" : "targets_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/TeamTargetAssignment"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "teams-controller" ]
      },
      "put" : {
        "operationId" : "setTargets_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "items" : {
                  "$ref" : "#/components/schemas/TeamTargetAssignment"
                },
                "type" : "array"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/TeamTargetAssignment"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "teams-controller" ]
      }
    },
...
    "/api/v1/users/{id}/targets" : {
      "get" : {
        "operationId" : "targets",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/UserTargetAssignment"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "users-controller" ]
      },
      "put" : {
        "operationId" : "setTargets",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "items" : {
                  "$ref" : "#/components/schemas/UserTargetAssignment"
                },
                "type" : "array"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/UserTargetAssignment"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "users-controller" ]
      }
    },
...
  },
  "security" : [ {
    "BearerAuth" : [ ]
  } ],
  "servers" : [ {
    "description" : "Current Vectispire Instance",
    "url" : "/"
  } ],
...
}